Get routing working with Vue template files
