+++ package.json
@@ -47,6 +47,7 @@
     "@feathersjs/express": "^1.2.7",
     "compression": "^1.7.3",
     "cors": "^2.8.5",
+    "css-loader": "^1.0.1",
     "feathers-mongoose": "^6.2.0",
     "helmet": "^3.15.0",
     "localstorage-memory": "^1.0.3",
@@ -70,6 +71,9 @@
     "nodemon": "^1.18.6",
     "request": "^2.88.0",
     "request-promise": "^4.2.2",
+    "vue-loader": "^15.4.2",
+    "vue-style-loader": "^4.1.2",
+    "vue-template-compiler": "^2.5.17",
     "webpack": "^4.26.1",
     "webpack-cli": "^3.1.2"
   }
