+++ src/client/components/Foo.vue
@@ -1,0 +1,9 @@
+<template>
+    <div class="vue-loaded">vue-loaded bitches</div>
+</template>
+
+<style>
+    .vue-loaded {
+        font-weight: bold;
+    }
+</style>
